% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDashboard.R
\name{runDashboard}
\alias{runDashboard}
\title{Launch Shiny Dashboard}
\usage{
runDashboard()
}
\description{
\code{runDashboard} launches the Shiny Dashboard displaying the data contained in this package
}
